+++ CHANGES.rst
@@ -62,3 +62,5 @@ the condition with speech-shaped noise only, the difference is smaller than 0
 .1%.
 - The py:func:`~pambox.speech.material.Material.set_level` function uses
 compensates for the reference sentence level using the correct sign.
+- Fix the py:func:`~pambox.utils.hilbert` definition that had been mangled in
+a merge. It is now the same as in `scipy.signal`.